+++ docs/changelog.md
@@ -16,6 +16,7 @@ category: Administration
 * feat: verify support for Jira 8.20.0 and latest enterprise patch releases
 * feat: verify support for Jira 8.20.1 and latest enterprise patch releases
 * test: improve test setup for displayed reference issues
+* feat: render issue description with the configured renderer 
 
 ### [21.10.0] - 2021-10-06